+++ src/lib/css-tag.ts
@@ -12,13 +12,20 @@ found at http://polymer.github.io/PATENTS.txt
 export const supportsAdoptingStyleSheets =
     ('adoptedStyleSheets' in Document.prototype);
 
+const constructionToken = Symbol();
+
 export class CSSResult {
 
   _styleSheet?: CSSStyleSheet|null;
 
   readonly cssText: string;
 
-  constructor(cssText: string) { this.cssText = cssText; }
+  constructor(cssText: string, safeToken: symbol) {
+    if (safeToken !== constructionToken) {
+      throw new Error('CSSResult is not constructable. Use `unsafeCss` or `css` instead.');
+    }
+    this.cssText = cssText;
+  }
 
   // Note, this is a getter so that it's lazy. In practice, this means
   // stylesheets are not created until the first element instance is made.
@@ -37,20 +44,38 @@ export class CSSResult {
   }
 }
 
+/**
+ * Wrap a value for interpolation in a css tagged template literal.
+ *
+ * This is unsafe because untrusted CSS text can be used to phone home
+ * or exfiltrate data to an attacker controlled site. Take care to only use
+ * this with trusted input.
+ */
+export const unsafeCss = (value: unknown) => {
+  return new CSSResult(String(value), constructionToken);
+};
+
 const textFromCSSResult = (value: CSSResult) => {
   if (value instanceof CSSResult) {
     return value.cssText;
   } else {
     throw new Error(
         `Value passed to 'css' function must be a 'css' function result: ${
-            value}.`);
+            value}. Use 'unsafeCss' to pass non-literal values, but
+            take care to ensure page security.` );
   }
 };
 
+/**
+ * Template tag which which can be used with LitElement's `style` property to
+ * set element styles. For security reasons, only literal string values may be
+ * used. To incorporate non-literal values `unsafeCss` may be used inside a
+ * template string part.
+ */
 export const css =
-    (strings: TemplateStringsArray, ...values: CSSResult[]): CSSResult => {
+    (strings: TemplateStringsArray, ...values: CSSResult[]) => {
       const cssText = values.reduce(
           (acc, v, idx) => acc + textFromCSSResult(v) + strings[idx + 1],
           strings[0]);
-      return new CSSResult(cssText);
+      return new CSSResult(cssText, constructionToken);
     };